Strictly Typed
This is what seems to be the shape of the JSON version, I got this from the README and formatted it a bit. Then I used that as a reference to write the types for the API itself.

I added all of the types with the tsconfig unedited, then just before this commit I unflagged these issues, since they are more related to code usage-related issues, and I usually only try to properly annotate those for certain once the codebase is fully typed and or in TS, because you can use more expressive nullish assertions there.

```json
{
  "tracks": [
    {
      "title": "Test",
      "length": 0,
      "params": [Object],
      "file": "http://sdfsdfsd.ua"
    },
    {
      "title": "Test2", "length": -1, "params": {}, "file": "http://url2"
    },
    {
      "title": "Test3", "length": 500, "params": {}, "file": "http://url3"
    }
  ],
  "header": {
    "param1": "val1",
    "param2": "val2"
  }
}
```