Should to Expect
Personally I like this test setup more because it explicitly uses parts of the testing library, without needing to modify the environment to add things to the global prototype for Object. I like the concept of how that works a lot, but I don't want to do something accidentally in the event of that.

https://www.chaijs.com/api/bdd/
https://stackoverflow.com/questions/21396524/what-is-the-difference-between-assert-expect-and-should-in-chai