Type Formatting
I like using interfaces for object shapes, I think it makes it easier to use recursive types, when need be. Merging is neat too, that's less common of a thing to worry about, but it is really nice. I think mapped types could be something you worry about as a usage type, rather than an implementation type. So maybe the user of the library would use the mapped type, and any edits they make to it could be derived upstream from an interface. Not that complex in this case to need something like that, but yeah! Super powerful :)