Formatting + Music Analysis (off topic)
Listening back again to The Thing That Knowledge Can't Eat, after listening to a review of it, and watching some classic Keneally videos with Marco Minnemann and Bryan Beller.

I'm starting to realize that even artists that are really broad in their sound do end up having to repeat themselves too, it's not unique to any of us, we all have to deal with it! I've been realizing this in myself lately, and it was making me feel bad. Celery actually kind of sounds like Wright or Wrong, and maybe even something from Boil That Dust Speck. Mercury sounded like Inkling. Maybe it's not that we are really good at doing a bunch of different things, but that we get really good at doing the things we know how to do. This feels like a new inspiring concept for me to think over!

Sorry this was a big write-up in not my own repo, usually I try not to do this outside of commits of projects on my own profile. It's like a nice place to write my thoughts out, since I'm writing here anyways :)