Symmetry Test
In this case, the file doesn't serialize symmetrically. So maybe I should add a test file that will be ensured to be serialized symmetrically, to validate that things are being written the same both ways every time. So this one isn't passing, but that's more of an intentional discovery. Now I know what the output of the project is, in correlation to what the test itself is supposed to account for.